requirements.txt
# List all Python packages needed
Flask
yfinance
mysql-connector-python
python-dotenv